fix: corrigir cadastro inicial no banco de dados
