+++ src/main/java/br/com/fiap/techchallenge/domain/service/CondutorService.java
@@ -1,15 +1,12 @@
 package br.com.fiap.techchallenge.domain.service;
 
 import br.com.fiap.techchallenge.domain.model.Condutor;
-import br.com.fiap.techchallenge.domain.model.Veiculo;
 import br.com.fiap.techchallenge.domain.repository.CondutorRepository;
-import jakarta.annotation.PostConstruct;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -47,23 +44,7 @@ public class CondutorService {
         return condutorRepository.findByNomeCompletoIgnoreCaseContaining(nomeCompleto);
     }
 
-    @PostConstruct()
-    private void criarDadosNoDB(){
-        List<Condutor> condutores = new ArrayList<>();
-        condutores.add(Condutor.builder()
-                .nomeCompleto("Ricardo Rafael Lara Gomes")
-                .email("ricardolara123@gmail.com")
-                .cpf("226.166.230-05")
-                .celular("(15) 99611-3333")
-                .build());
-
-        condutores.add(Condutor.builder()
-                .nomeCompleto("Fulano da Silva")
-                .email("fulano.dasilva@gmail.com")
-                .cpf("005.322.150-87")
-                .celular("(15) 88622-3333")
-                .build());
-
-        condutorRepository.saveAll(condutores);
+    public void saveAll(List<Condutor> lista){
+        condutorRepository.saveAll(lista);
     }
 }
